[류현종] classOpenApplyDAO,Professor,Student,Department - 삭제검증메소드추가[과목이 있을때] deleteDepartment,deleteSubject-DAO추가

+++ lms/src/main/webapp/admin/department/deleteDepartment.jsp
@@ -11,14 +11,24 @@
 	String department = request.getParameter("department");
 	System.out.println(department + "<-- deleteDepartment.jsp department");
 
-	//DAO호출
-	int deleteDepartment = DepartmentDAO.deleteDepartment(department);
 	
-	if(deleteDepartment==1){ //삭제 성공
-		System.out.println("학과가 삭제되었습니다");
-		response.sendRedirect("/lms/admin/department/departmentList.jsp");
-	}else{//삭제 실패
-		System.out.println("학과가 삭제되지 않았습니다");
+	int selectStudent = StudentDAO.selectStudent(department);
+	int selectProfessor = ProfessorDAO.selectProfessor(department);
+	//학과 삭제할때 검색
+	if(selectStudent==0&&selectProfessor==0){
+		//DAO호출
+		int deleteDepartment = DepartmentDAO.deleteDepartment(department);
+		//학생 찾기 (?)
+		
+		if(deleteDepartment==1){ //삭제 성공
+			System.out.println("학과가 삭제되었습니다");
+			response.sendRedirect("/lms/admin/department/departmentList.jsp");
+		}else{//삭제 실패
+			System.out.println("학과가 삭제되지 않았습니다");
+			response.sendRedirect("/lms/admin/department/departmentList.jsp");
+		}
+	}else{
+		System.out.println("교수 또는 학생이 존재합니다");
 		response.sendRedirect("/lms/admin/department/departmentList.jsp");
 	}
 %>
